Moving the entry sorter to the model as it is required by the BibtexDatabase.
