+++ src/main/java/net/sf/jabref/collab/ChangeScanner.java
@@ -39,7 +39,7 @@ import net.sf.jabref.gui.JabRefFrame;
 import net.sf.jabref.importer.OpenDatabaseAction;
 import net.sf.jabref.importer.ParserResult;
 import net.sf.jabref.bibtex.DuplicateCheck;
-import net.sf.jabref.bibtex.EntrySorter;
+import net.sf.jabref.model.database.EntrySorter;
 import net.sf.jabref.bibtex.comparator.EntryComparator;
 import net.sf.jabref.logic.l10n.Localization;
 import net.sf.jabref.model.database.BibtexDatabase;
