+++ src/main/java/net/sf/jabref/model/database/EntrySorter.java
@@ -13,11 +13,9 @@
     with this program; if not, write to the Free Software Foundation, Inc.,
     51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
 */
-package net.sf.jabref.bibtex;
+package net.sf.jabref.model.database;
 
 import net.sf.jabref.model.entry.BibtexEntry;
-import net.sf.jabref.model.database.DatabaseChangeEvent;
-import net.sf.jabref.model.database.DatabaseChangeListener;
 
 import java.util.*;
